requirements.txt
asgiref==3.4.1
boto3==1.17.104
colorama==0.4.4
commonmark==0.9.1
crispy-tailwind==0.5.0
dj-database-url==0.5.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-storages==1.11.1
django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
faker==8.9.0
pillow==8.3.0
psycopg2-binary==2.9.1
psycopg2==2.9.1
pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==1.0.1
pytz==2021.1
rich==10.4.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
